#14 number fields now in separate sub-package
also version is now 0.1
