+++ superfields/src/main/java/org/vaadin/miki/superfields/numbers/SuperBigDecimalField.java
@@ -1,4 +1,4 @@
-package org.vaadin.miki.superfields;
+package org.vaadin.miki.superfields.numbers;
 
 import com.vaadin.flow.component.Tag;
