+++ superfields/src/main/java/org/vaadin/miki/superfields/numbers/SuperLongField.java
@@ -1,4 +1,4 @@
-package org.vaadin.miki.superfields;
+package org.vaadin.miki.superfields.numbers;
 
 import com.vaadin.flow.component.Tag;
 
@@ -7,18 +7,18 @@ import java.text.ParseException;
 import java.util.Locale;
 
 /**
- * A field for {@link Integer} values that are properly formatted.
+ * A field for {@link Long} values that are properly formatted.
  *
  * @author miki
  * @since 2020-04-07
  */
-@Tag("super-integer-field")
-public class SuperIntegerField extends AbstractSuperNumberField<Integer> {
+@Tag("super-long-field")
+public class SuperLongField extends AbstractSuperNumberField<Long> {
 
     /**
      * Constructs the field with zero as default value, default {@link Locale} and an empty label.
      */
-    public SuperIntegerField() {
+    public SuperLongField() {
         this("");
     }
 
@@ -26,7 +26,7 @@ public class SuperIntegerField extends AbstractSuperNumberField<Integer> {
      * Constructs the field with zero as default value and an empty label.
      * @param locale Locale to use for formatting.
      */
-    public SuperIntegerField(Locale locale) {
+    public SuperLongField(Locale locale) {
         this("", locale);
     }
 
@@ -34,7 +34,7 @@ public class SuperIntegerField extends AbstractSuperNumberField<Integer> {
      * Constructs the field with zero as default value and default {@link Locale}.
      * @param label Label that accompanies the field.
      */
-    public SuperIntegerField(String label) {
+    public SuperLongField(String label) {
         this(label, Locale.getDefault());
     }
 
@@ -43,7 +43,7 @@ public class SuperIntegerField extends AbstractSuperNumberField<Integer> {
      * @param label Label that accompanies the field.
      * @param locale Locale to use for formatting.
      */
-    public SuperIntegerField(String label, Locale locale) {
+    public SuperLongField(String label, Locale locale) {
         this(0, label, locale);
     }
 
@@ -53,12 +53,12 @@ public class SuperIntegerField extends AbstractSuperNumberField<Integer> {
      * @param label Label that accompanies the field.
      * @param locale Locale to use for formatting.
      */
-    public SuperIntegerField(int defaultValue, String label, Locale locale) {
+    public SuperLongField(long defaultValue, String label, Locale locale) {
         super(defaultValue, d -> d < 0, Math::abs, label, locale, 0);
     }
 
     @Override
-    protected Integer parseRawValue(String rawValue, DecimalFormat format) throws ParseException {
-        return format.parse(rawValue).intValue();
+    protected Long parseRawValue(String rawValue, DecimalFormat format) throws ParseException {
+        return format.parse(rawValue).longValue();
     }
 }
