+++ superfields/src/test/java/org/vaadin/miki/superfields/numbers/SuperIntegerFieldTest.java
@@ -1,4 +1,4 @@
-package org.vaadin.miki.superfields;
+package org.vaadin.miki.superfields.numbers;
 
 import java.util.Locale;
